requirements.txt
torch
torchvision
pillow
streamlit
opencv-python
numpy<2
plotly
pandas
matplotlib